Camera WIP, read description
Camera seems to clip below the ground when too near to it (i.e. rat mesh is small)
If we increase rat mesh height, say to 20, the bug disappears.
Seems to be that it clips below somehow, and then due to physics, can never get back up. (Just a guess)